Refactor: Created packages for views, models, viewmodels, repositories and adapters
-The classes have been split into their respective packages
