+++ app/src/main/java/com/droidhats/campuscompass/models/Calendar.kt
@@ -1,4 +1,4 @@
-package com.droidhats.campuscompass
+package com.droidhats.campuscompass.models
 
 import android.provider.CalendarContract
