+++ app/src/main/java/com/droidhats/campuscompass/viewmodels/MapViewModel.kt
@@ -1,4 +1,4 @@
-package com.droidhats.campuscompass
+package com.droidhats.campuscompass.viewmodels
 
 import androidx.lifecycle.ViewModel
