+++ app/src/main/res/layout/calendar_fragment.xml
@@ -24,7 +24,7 @@
         android:layout_marginLeft="16dp"
         android:layout_marginRight="16dp"
         app:layoutManager="LinearLayoutManager"
-        tools:context=".CalendarFragment"
+        tools:context=".views.CalendarFragment"
         tools:listitem="@layout/calendar_recycler_item" />
 
 </LinearLayout>
