+++ app/src/main/res/layout/map_fragment.xml
@@ -7,7 +7,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/coordinate_layout"
-    tools:context="MapFragment">
+    tools:context=".views.MapFragment">
 
     <fragment
         android:id="@+id/map"
